Update for mythtv protocol version 18 fixes free space call.  Only provides free space on master backend, not slaves might fix later to add up across all slaves, but at least this gets it working again.

+++ src/mythtv.c
@@ -715,7 +715,7 @@ int
 mythtv_update(mvp_widget_t *widget)
 {
 	char buf[64];
-	int total, used;
+	long long total, used;
 
 	if (mythtv_livetv == 0) {
 		if (mythtv_recdir)
